Move ElasticsearchService into the shared library
