+++ search/src/main/scala/uk/ac/wellcome/platform/api/search/rest/ImagesController.scala
@@ -5,10 +5,7 @@ import com.sksamuel.elastic4s.Index
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport
 import uk.ac.wellcome.api.display.models.Implicits._
 import uk.ac.wellcome.platform.api.search.models.{QueryConfig, SimilarityMetric}
-import uk.ac.wellcome.platform.api.search.services.{
-  ElasticsearchService,
-  ImagesService
-}
+import uk.ac.wellcome.platform.api.search.services.ImagesService
 import cats.implicits._
 import uk.ac.wellcome.api.display.models.{
   DisplayImage,
@@ -18,6 +15,7 @@ import uk.ac.wellcome.api.display.models.{
 import uk.ac.wellcome.Tracing
 import uk.ac.wellcome.platform.api.rest.CustomDirectives
 import uk.ac.wellcome.platform.api.models.ApiConfig
+import weco.api.search.elasticsearch.ElasticsearchService
 import weco.catalogue.internal_model.identifiers.CanonicalId
 import weco.http.models.ContextResponse
