+++ search/src/main/scala/uk/ac/wellcome/platform/api/search/services/WorksService.scala
@@ -4,6 +4,7 @@ import com.sksamuel.elastic4s.requests.searches.SearchResponse
 import io.circe.Decoder
 import uk.ac.wellcome.models.Implicits
 import uk.ac.wellcome.platform.api.search.models._
+import weco.api.search.elasticsearch.ElasticsearchService
 import weco.catalogue.internal_model.work.Work
 import weco.catalogue.internal_model.work.WorkState.Indexed
